+++ packages/proper-changelog/src/types.ts
@@ -1,4 +1,5 @@
 import type { components } from '@octokit/openapi-types';
+import type { CommanderError, OutputConfiguration } from 'commander';
 
 /** A GitHub release as returned by the REST API (`GET /repos/{owner}/{repo}/releases`). */
 export type GitHubRelease = components['schemas']['release'];
@@ -10,7 +11,7 @@ export interface RepoId {
 }
 
 /** Options as returned by `program.parse().opts()`. */
-export interface RawCliOptions {
+export interface CliOptions {
   /** Repository to read releases from. */
   repo?: RepoId;
   /** npm package name the repo was resolved from, if any (used for the changelog heading/filename). */
@@ -26,10 +27,11 @@ export interface RawCliOptions {
   /** Maximum number of releases to include. */
   limit?: number;
   /**
-   * Filter releases by tag name. A plain string matches tags that contain it (case-insensitive);
-   * a value wrapped in slashes (e.g. `/^v1\./i`) is treated as a regular expression.
+   * Filter releases by tag name. A string matches tags that contain it (case-insensitive);
+   * a `RegExp` matches tags that satisfy it. (The CLI converts a value wrapped in slashes,
+   * e.g. `/^v1\./i`, into a `RegExp`.)
    */
-  filter?: string;
+  filter?: string | RegExp;
   /** Only include releases published after this date. */
   since?: Date;
   /** Write output to this file */
@@ -39,7 +41,19 @@ export interface RawCliOptions {
 }
 
 /** Options controlling changelog generation. */
-export type ProperChangelogOptions = Required<Pick<RawCliOptions, 'repo'>> & RawCliOptions;
+export type ProperChangelogOptions = Required<Pick<CliOptions, 'repo'>> & CliOptions;
+
+export interface CliContext {
+  argv: string[];
+  env: NodeJS.ProcessEnv;
+  /** Commander error handler */
+  exitOverride?: (err: CommanderError) => never | void;
+  /** Commander error logging handler */
+  writeErr?: OutputConfiguration['writeErr'];
+  log: (message: string) => void;
+  warn: (message: string) => void;
+  writeFile: (file: string, content: string) => void;
+}
 
 /** Throw this to indicate an expected error (stack won't be logged) */
 export class ChangelogError extends Error {
